backend/pyproject.toml
[tool.poetry]
name = "indors-test-task"
version = "0.1.0"
description = ""
authors = ["Sergey <one6sergey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
ruff = "^0.6.7"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
djoser = "^2.2.3"
django-cors-headers = "^4.4.0"
channels = "^4.1.0"
daphne = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
